Refactor newly added Range support
Minor changes applied as a follow-up to PR #178:

- concentrate new Range functionality in dedicated files
  - this reflects similar work underway for GH-175
- update README
  - mention new Range support
  - link to dependent projects (including one from contributor
    of new functionality)
- doc comment tweaks
- misc changes to satisfy current CI linters

Credit for original work:

- @infraweavers (https://github.com/infraweavers)
- https://github.com/atc0005/go-nagios/pull/178

refs GH-176
refs GH-178
